requirements.txt
streamlit
pandas
numpy
python-docx
PyMuPDF
Pillow
openpyxl
rapidfuzz
pdfplumber